better pyi script

+++ python.sh
@@ -3,11 +3,28 @@ alias py_env_activate="source ./.pyenv/env/bin/activate"
 
 alias py_test="pytest"
 
-function pyi(){
-   pip3 install -I -r requirements.txt
-   for d in src/functions/*/ ; do
-   pip3 install -I -r "$d"/requirements.txt
-   done
+function pyi() {
+    if [ -f setup.py ]; then
+        pip3 install -e '.[dev]'
+    fi
+
+    if [ -f requirements.txt ]; then
+        pip3 install -I -r requirements.txt
+    fi
+
+    if [ -f requirements-dev.txt ]; then
+        pip3 install -I -r requirements-dev.txt
+    fi
+
+    for d in src/functions/*/; do
+        if [ -f "$d"/requirements.txt ]; then
+            pip3 install -I -r "$d"/requirements.txt
+        fi
+        if [ -f "$d"/requirements-dev.txt ]; then
+            pip3 install -I -r "$d"/requirements-dev.txt
+        fi
+    done
+
 }
 
 if [ -f ./.pyenv/env/bin/activate ]; then
